+++ app/controllers/doorkeeper/openid_connect/userinfo_controller.rb
@@ -6,7 +6,8 @@ module Doorkeeper
       before_action -> { doorkeeper_authorize! :openid }
 
       def show
-        render json: Doorkeeper::OpenidConnect::UserInfo.new(doorkeeper_token), status: :ok
+        render json: Doorkeeper::OpenidConnect.configuration.user_info_model.new(doorkeeper_token),
+               status: :ok
       end
     end
   end